fix(mpc-attestation): introduce local_verify helper for off-chain consumers
The first commit of this PR (96773d3a) split `Attestation::verify` into
`extract_dcap_inputs` (Promise path) + `finish_verify` (callback path)
to remove `dcap-qvl` from the contract WASM. Off-chain consumers
(`tee-authority`, `attestation-cli`, `mpc-node`, integration tests)
still needed a one-call local verifier.

This commit:

- Adds `mpc_attestation::local_verify::local_verify`, gated behind a
  new `local-verify` feature on `mpc-attestation`. It chains
  `DstackVerify::verify` (dcap-qvl) and `finish_verify` (post-DCAP)
  into a single call.
- Switches `tee-authority`, `attestation-cli`, `mpc-node`, and the
  `mpc-attestation` integration tests to call `local_verify`,
  removing duplicate two-step helpers.
- Drops `attestation` as a direct dependency of `attestation-cli`
  and `tee-authority`; both now import everything they need through
  `mpc-attestation` and `attestation-types`.
- Keeps a test-only `TeeState::add_participant` wrapper so existing
  contract unit tests don't have to be rewritten. The wrapper
  internally calls `Attestation::finish_verify` against a dummy
  `VerifiedReport`, which is fine for `Mock` attestations (the only
  variant tests use directly).
