+++ crates/attestation-cli/Cargo.toml
@@ -10,7 +10,7 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = { workspace = true }
-attestation = { workspace = true }
+attestation-types = { workspace = true }
 bs58 = { workspace = true }
 clap = { workspace = true }
 mpc-attestation = { workspace = true, features = ["local-verify"] }
