+++ crates/mpc-attestation/tests/test_attestation_verification.rs
@@ -1,183 +1,16 @@
 use assert_matches::assert_matches;
-use attestation::attestation::DstackVerify as _;
 use attestation_types::measurements::{ExpectedMeasurements, Measurements};
 use attestation_types::verify_post_dcap::VerificationError;
 use mpc_attestation::attestation::{
     Attestation, DEFAULT_EXPIRATION_DURATION_SECONDS, MockAttestation, VerifiedAttestation,
     default_measurements,
 };
+use mpc_attestation::local_verify::local_verify;
 use mpc_attestation::report_data::{ReportData, ReportDataV1};
-use mpc_primitives::hash::{LauncherDockerComposeHash, NodeImageHash};
 use test_utils::attestation::{
     VALID_ATTESTATION_TIMESTAMP, account_key, image_digest, launcher_compose_digest,
     mock_dstack_attestation, p2p_tls_key,
 };
-
-/// Off-chain convenience: runs the full dcap-qvl + post-DCAP path in
-/// one call. Mirrors what the deleted `Attestation::verify` method used
-/// to do. Lives here because `mpc-attestation` no longer depends on
-/// the `dcap-qvl`-using `attestation` crate.
-fn local_verify(
-    attestation: &Attestation,
-    expected_report_data: ReportData,
-    timestamp_seconds: u64,
-    allowed_mpc_docker_image_hashes: &[NodeImageHash],
-    allowed_launcher_docker_compose_hashes: &[LauncherDockerComposeHash],
-    accepted_measurements: &[ExpectedMeasurements],
-) -> Result<VerifiedAttestation, VerificationError> {
-    let verified_report = match attestation {
-        Attestation::Dstack(dstack) => {
-            let measurements = dstack.verify(
-                expected_report_data.clone(),
-                timestamp_seconds,
-                accepted_measurements,
-            )?;
-            // `DstackVerify::verify` already ran the post-DCAP checks
-            // for us, including measurement matching. We still need a
-            // `VerifiedReport` to pass into `finish_verify` (which
-            // re-runs the same checks) — build it by re-calling the
-            // local conversion. This path is off-chain and the double
-            // work is negligible.
-            //
-            // The simpler alternative would be to factor a `verify_raw`
-            // out of `attestation` that returns the mirror report
-            // directly. For test code, this is fine; production callers
-            // (the `mpc-contract` Promise callback) get the mirror
-            // directly from the verifier contract.
-            let _ = measurements;
-            local_dcap_verify(dstack, timestamp_seconds)?
-        }
-        Attestation::Mock(_) => zero_verified_report(),
-    };
-
-    attestation.finish_verify(
-        &verified_report,
-        expected_report_data,
-        allowed_mpc_docker_image_hashes,
-        allowed_launcher_docker_compose_hashes,
-        accepted_measurements,
-        timestamp_seconds,
-    )
-}
-
-/// Calls `dcap_qvl::verify::verify` against a `DstackAttestation` and
-/// returns the result as the Borsh mirror.
-fn local_dcap_verify(
-    dstack: &mpc_attestation::attestation::DstackAttestation,
-    timestamp_seconds: u64,
-) -> Result<tee_verifier_interface::VerifiedReport, VerificationError> {
-    let collateral = attestation::collateral_to_dcap(dstack.collateral.clone());
-    let verification_result =
-        dcap_qvl::verify::verify(&dstack.quote, &collateral, timestamp_seconds)
-            .map_err(|e| VerificationError::DcapVerification(e.to_string()))?;
-    Ok(dcap_to_mirror(verification_result))
-}
-
-fn dcap_to_mirror(
-    value: dcap_qvl::verify::VerifiedReport,
-) -> tee_verifier_interface::VerifiedReport {
-    tee_verifier_interface::VerifiedReport {
-        status: value.status,
-        advisory_ids: value.advisory_ids,
-        report: match value.report {
-            dcap_qvl::quote::Report::SgxEnclave(_) => unreachable!("SGX not used in tests"),
-            dcap_qvl::quote::Report::TD10(r) => {
-                tee_verifier_interface::Report::TD10(td10_to_mirror(r))
-            }
-            dcap_qvl::quote::Report::TD15(r) => {
-                tee_verifier_interface::Report::TD15(tee_verifier_interface::TDReport15 {
-                    base: td10_to_mirror(r.base),
-                    tee_tcb_svn2: r.tee_tcb_svn2,
-                    mr_service_td: r.mr_service_td,
-                })
-            }
-        },
-        ppid: value.ppid,
-        qe_status: tcb_status_with_advisory_to_mirror(value.qe_status),
-        platform_status: tcb_status_with_advisory_to_mirror(value.platform_status),
-    }
-}
-
-fn td10_to_mirror(value: dcap_qvl::quote::TDReport10) -> tee_verifier_interface::TDReport10 {
-    tee_verifier_interface::TDReport10 {
-        tee_tcb_svn: value.tee_tcb_svn,
-        mr_seam: value.mr_seam,
-        mr_signer_seam: value.mr_signer_seam,
-        seam_attributes: value.seam_attributes,
-        td_attributes: value.td_attributes,
-        xfam: value.xfam,
-        mr_td: value.mr_td,
-        mr_config_id: value.mr_config_id,
-        mr_owner: value.mr_owner,
-        mr_owner_config: value.mr_owner_config,
-        rt_mr0: value.rt_mr0,
-        rt_mr1: value.rt_mr1,
-        rt_mr2: value.rt_mr2,
-        rt_mr3: value.rt_mr3,
-        report_data: value.report_data,
-    }
-}
-
-fn tcb_status_with_advisory_to_mirror(
-    value: dcap_qvl::tcb_info::TcbStatusWithAdvisory,
-) -> tee_verifier_interface::TcbStatusWithAdvisory {
-    tee_verifier_interface::TcbStatusWithAdvisory {
-        status: match value.status {
-            dcap_qvl::tcb_info::TcbStatus::UpToDate => tee_verifier_interface::TcbStatus::UpToDate,
-            dcap_qvl::tcb_info::TcbStatus::OutOfDateConfigurationNeeded => {
-                tee_verifier_interface::TcbStatus::OutOfDateConfigurationNeeded
-            }
-            dcap_qvl::tcb_info::TcbStatus::OutOfDate => {
-                tee_verifier_interface::TcbStatus::OutOfDate
-            }
-            dcap_qvl::tcb_info::TcbStatus::ConfigurationAndSWHardeningNeeded => {
-                tee_verifier_interface::TcbStatus::ConfigurationAndSWHardeningNeeded
-            }
-            dcap_qvl::tcb_info::TcbStatus::ConfigurationNeeded => {
-                tee_verifier_interface::TcbStatus::ConfigurationNeeded
-            }
-            dcap_qvl::tcb_info::TcbStatus::SWHardeningNeeded => {
-                tee_verifier_interface::TcbStatus::SWHardeningNeeded
-            }
-            dcap_qvl::tcb_info::TcbStatus::Revoked => tee_verifier_interface::TcbStatus::Revoked,
-        },
-        advisory_ids: value.advisory_ids,
-    }
-}
-
-/// Zero-valued `VerifiedReport` for `Mock` paths that ignore it.
-fn zero_verified_report() -> tee_verifier_interface::VerifiedReport {
-    tee_verifier_interface::VerifiedReport {
-        status: String::new(),
-        advisory_ids: Vec::new(),
-        report: tee_verifier_interface::Report::TD10(tee_verifier_interface::TDReport10 {
-            tee_tcb_svn: [0; 16],
-            mr_seam: [0; 48],
-            mr_signer_seam: [0; 48],
-            seam_attributes: [0; 8],
-            td_attributes: [0; 8],
-            xfam: [0; 8],
-            mr_td: [0; 48],
-            mr_config_id: [0; 48],
-            mr_owner: [0; 48],
-            mr_owner_config: [0; 48],
-            rt_mr0: [0; 48],
-            rt_mr1: [0; 48],
-            rt_mr2: [0; 48],
-            rt_mr3: [0; 48],
-            report_data: [0; 64],
-        }),
-        ppid: Vec::new(),
-        qe_status: tee_verifier_interface::TcbStatusWithAdvisory {
-            status: tee_verifier_interface::TcbStatus::UpToDate,
-            advisory_ids: Vec::new(),
-        },
-        platform_status: tee_verifier_interface::TcbStatusWithAdvisory {
-            status: tee_verifier_interface::TcbStatus::UpToDate,
-            advisory_ids: Vec::new(),
-        },
-    }
-}
 
 #[test]
 fn valid_mock_attestation_succeeds_verification() {
@@ -189,7 +22,14 @@ fn valid_mock_attestation_succeeds_verification() {
     let report_data = ReportData::V1(ReportDataV1::new(tls_key, account_key));
 
     assert_matches!(
-        local_verify(&valid_attestation, report_data, timestamp_s, &[], &[], &[]),
+        local_verify(
+            &valid_attestation,
+            report_data.into(),
+            timestamp_s,
+            &[],
+            &[],
+            &[]
+        ),
         Ok(VerifiedAttestation::Mock(MockAttestation::Valid))
     );
 }
@@ -204,7 +44,14 @@ fn invalid_mock_attestation_fails_verification() {
     let report_data = ReportData::V1(ReportDataV1::new(tls_key, account_key));
 
     assert_matches!(
-        local_verify(&valid_attestation, report_data, timestamp_s, &[], &[], &[]),
+        local_verify(
+            &valid_attestation,
+            report_data.into(),
+            timestamp_s,
+            &[],
+            &[],
+            &[]
+        ),
         Err(VerificationError::InvalidMockAttestation)
     );
 }
@@ -222,7 +69,7 @@ fn validated_dstack_attestation_can_be_reverified() {
 
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         timestamp_s,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
@@ -255,7 +102,7 @@ fn validated_dstack_attestation_fails_reverification_when_expired() {
 
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         timestamp_s,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
@@ -285,7 +132,7 @@ fn validated_mock_attestation_passes_reverification() {
     let account_key = account_key();
     let report_data: ReportData = ReportDataV1::new(tls_key, account_key).into();
 
-    let validated = local_verify(&valid_attestation, report_data, 0, &[], &[], &[])
+    let validated = local_verify(&valid_attestation, report_data.into(), 0, &[], &[], &[])
         .expect("Initial verification failed");
 
     // Mock should generally pass re-verify
@@ -306,7 +153,7 @@ fn validated_dstack_attestation_fails_reverification_with_rotated_hashes() {
     // 1. Initial verify succeeds with the "old" allowed list
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         creation_time,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
@@ -344,7 +191,7 @@ fn validated_dstack_attestation_fails_reverification_with_removed_measurements()
 
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         creation_time,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
@@ -387,7 +234,7 @@ fn validated_dstack_attestation_fails_reverification_with_empty_measurements() {
 
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         creation_time,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
@@ -420,7 +267,7 @@ fn validated_dstack_attestation_passes_reverification_with_superset_measurements
 
     let validated = local_verify(
         &attestation,
-        report_data,
+        report_data.into(),
         creation_time,
         &allowed_mpc_hashes,
         &allowed_launcher_hashes,
