+++ crates/node/Cargo.toml
@@ -33,7 +33,7 @@ itertools = { workspace = true }
 k256 = { workspace = true }
 launcher-interface = { workspace = true }
 lru = { workspace = true }
-mpc-attestation = { workspace = true }
+mpc-attestation = { workspace = true, features = ["local-verify"] }
 mpc-node-config = { workspace = true }
 mpc-primitives = { workspace = true }
 mpc-tls = { workspace = true }
