+++ crates/tee-authority/Cargo.toml
@@ -9,7 +9,6 @@ external-services-tests = []
 
 [dependencies]
 anyhow = { workspace = true }
-attestation = { workspace = true }
 backon = { workspace = true }
 dcap-qvl = { workspace = true, features = ["report"] }
 derive_more = { workspace = true }
@@ -32,7 +31,6 @@ x509-parser = { workspace = true }
 assert_matches = { workspace = true }
 hex = { workspace = true }
 rstest = { workspace = true }
-tee-verifier-interface = { workspace = true }
 test-utils = { workspace = true }
 tokio = { workspace = true, features = ["test-util"] }
